WOO-501
When customers are clicking "purchase button" (intercept), the order is tagged with a purchase start time.
When customers are landing or completing the order is tagged with a purcase end time.
Result: If customers are leaving signing or not fulfulling the order, there will be no end time stamp and therefore easier to track down.