WOO-593
The code logics in this plugin prevents most of the use of "time throttling", at least in the admin parts.
However, in the checkout we can adjust the time on fly for which we use to wait for Resurs responses. If there are timeouts after the first 12 (default) seconds in the request the plugin will push down the next request during a minute to 3 (default, but configurable) seconds.